My bad!

+++ quotes_app/quotes.py
@@ -85,11 +85,13 @@ def get_quotes():
     for item in data:
         item["_id"] = str(item["_id"])
         item["object"] = ObjectId(item["_id"])
+        item["comments"] = list(comments_db.comments_collection.find({"quote_id": item["_id"]}))[::-1]
     # display the data
     html = render_template(
         "quotes.html",
-        data=data,
+        data=data[::-1],
         user=user,
+        page=page_display,
     )
     response = make_response(html)
     response.set_cookie("session_id", session_id)
@@ -180,6 +182,45 @@ def get_logout():
     response.delete_cookie("session_id")
     return response
 
+@app.route("/comment", methods=["POST"])
+def post_comment():
+    session_id = request.cookies.get("session_id", None)
+    if not session_id:
+        response = redirect("/login")
+        return response
+    # open the session collection
+    session_collection = session_db.session_collection
+    # get the data for this session
+    session_data = list(session_collection.find({"session_id": session_id}))
+    if len(session_data) == 0:
+        response = redirect("/logout")
+        return response
+    assert len(session_data) == 1
+    session_data = session_data[0]
+    # get some information from the session
+    user = session_data.get("user", "unknown user")
+    text = request.form.get("text", "")
+    quote_id = request.form.get("quoteID", "")
+    if text != "" and quote_id != "":
+        #if quote owner
+        quote_collection = quotes_db.quotes_collection
+        quote = quote_collection.find_one({"_id": quote_id})
+        # open the comments collection
+        comments_collection = comments_db.comments_collection
+        # insert the comment
+        print(quote)
+        if quote["allow_comments"] == "true":
+            comment_data = {"owner": user, "text": text, "quote_id": quote_id}
+            comments_collection.insert_one(comment_data)
+            print(comment_data)
+        else:
+            print("Comments not allowed for this quote")
+    # usually do a redirect('....')
+    page = request.form.get("page", "")
+    if page == "My Quotes":
+        return redirect("/myquotes")
+    else:
+        return redirect("/quotes")
 
 @app.route("/add", methods=["POST"])
 def post_add():
@@ -200,14 +241,20 @@ def post_add():
     user = session_data.get("user", "unknown user")
     text = request.form.get("text", "")
     author = request.form.get("author", "")
-    if text != "" and author != "":
+    access = request.form.get("access", "")
+    allow_comments = request.form.get("allowComments", "false")
+    if text != "" and author != "" and access != "":
         # open the quotes collection
         quotes_collection = quotes_db.quotes_collection
         # insert the quote
-        quote_data = {"owner": user, "text": text, "author": author}
+        quote_data = {"owner": user, "text": text, "author": author, "access": access, "allow_comments": allow_comments}
+        print(quote_data)
         quotes_collection.insert_one(quote_data)
     # usually do a redirect('....')
-    return redirect("/quotes")
+    if access == "Private":
+        return redirect("/myquotes")
+    else:
+        return redirect("/quotes")
 
 @app.route("/edit", methods=["POST"])
 def post_edit():
@@ -218,29 +265,79 @@ def post_edit():
     _id = request.form.get("_id", None)
     text = request.form.get("text", "")
     author = request.form.get("author", "")
+    access = request.form.get("access", "")
+    allow_comments = request.form.get("allowComments", "")
     if _id:
+        session_collection = session_db.session_collection
+        session_data = list(session_collection.find({"session_id": session_id}))
+        session_data = session_data[0]
+        user = session_data.get("user", "unknown user")
         # open the quotes collection
         quotes_collection = quotes_db.quotes_collection
+        quote = quotes_db.quotes_collection.find_one({"_id": ObjectId(_id)})
         # update the values in this particular record
-        values = {"$set": {"text": text, "author": author}}
-        data = quotes_collection.update_one({"_id": ObjectId(_id)}, values)
+        if user == quote["owner"]:
+            values = {"$set": {"text": text, "author": author, "access": access, "allow_comments": allow_comments}}
+            print(values)
+            data = quotes_collection.update_one({"_id": ObjectId(_id)}, values)
     # do a redirect('....')
-    return redirect("/quotes")
+    if access == "Private":
+        return redirect("/myquotes")
+    else:
+        return redirect("/quotes")
 
 
 @app.route("/delete", methods=["GET"])
 @app.route("/delete/<id>", methods=["GET"])
 def get_delete(id=None):
     session_id = request.cookies.get("session_id", None)
+    print(request)
     if not session_id:
         response = redirect("/login")
         return response
     if id:
+        session_collection = session_db.session_collection
+        session_data = list(session_collection.find({"session_id": session_id}))
+        session_data = session_data[0]
+        user = session_data.get("user", "unknown user")
         # open the quotes collection
         quotes_collection = quotes_db.quotes_collection
+        comment_collection = comments_db.comments_collection
+        quote = quotes_db.quotes_collection.find_one({"_id": ObjectId(id)})
+        print(quote)
+        if user == quote["owner"]:
+            comment_collection.delete_many({"quote_id": ObjectId(id)})
+            quotes_collection.delete_one({"_id": ObjectId(id)})
         # delete the item
-        quotes_collection.delete_one({"_id": ObjectId(id)})
+        # quotes_collection.delete_one({"_id": ObjectId(id)})
+    # return to the quotes page
+    return redirect("/myquotes")
+
+@app.route("/delete/comment/<id>", methods=["GET"])
+def get_delete_comment(id=None):
+    session_id = request.cookies.get("session_id", None)
+    if not session_id:
+        response = redirect("/login")
+        return response
+    if id:
+        # get some information from the session
+        session_collection = session_db.session_collection
+        session_data = list(session_collection.find({"session_id": session_id}))
+        session_data = session_data[0]
+        user = session_data.get("user", "unknown user")
+
+        #if comment owner
+        comment_collection = comments_db.comments_collection
+        comment = comment_collection.find_one({"_id": ObjectId(id)})
+        if user == comment["owner"]:
+            comment_collection.delete_one({"_id": ObjectId(id)})
+            print("Comment deleted by comment owner")
+
+        #if quote owner
+        quote_collection = quotes_db.quotes_collection
+        quote = quote_collection.find_one({"_id": comment["quote_id"]})
+        if user == quote["owner"]:
+            comment_collection.delete_one({"_id": ObjectId(id)})
+            print("Comment deleted by quote owner")
     # return to the quotes page
     return redirect("/quotes")
-
-app.run(port=5000, debug=True)
